backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
arabic-reshaper==3.0.0
async-timeout==4.0.3
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cohere==5.13.2
cryptography==44.0.0
dataclasses-json==0.6.7
exceptiongroup==1.2.2
fastavro==1.9.7
filelock==3.16.1
Flask==2.2.5
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-huggingface==0.0.3
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.4
langsmith==0.1.147
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
orjson==3.10.12
packaging==24.2
parameterized==0.9.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyMuPDF==1.24.11
pypdfium2==4.30.1
python-bidi==0.6.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.1
transformers==4.46.3
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
